Move StartMenu to scenes package
This cleans up the Core module.
