+++ Core/src/main/resources/com/github/sef24sp4/core/scenes/start-menu.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.github.sef24sp4.core.StartMenu">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.github.sef24sp4.core.scenes.StartMenu">
    <children>
       <Button fx:id="startButton" alignment="CENTER" layoutX="240.0" layoutY="166.0" mnemonicParsing="false" onAction="#startGame" text="Start">
          <font>
